KAN-230 refactor: commonSetup 생성 및 적용
